Simplified LightningView with externalized XXXClients
