+++ app/src/main/java/acr/browser/lightning/app/AppComponent.java
@@ -8,7 +8,8 @@ import javax.inject.Singleton;
 
 import acr.browser.lightning.activity.BrowserActivity;
 import acr.browser.lightning.constant.BookmarkPage;
-import acr.browser.lightning.dialog.BookmarksDialogBuilder;
+import acr.browser.lightning.database.HistoryDatabase;
+import acr.browser.lightning.dialog.LightningDialogBuilder;
 import acr.browser.lightning.fragment.BookmarkSettingsFragment;
 import acr.browser.lightning.fragment.BookmarksFragment;
 import acr.browser.lightning.fragment.LightningPreferenceFragment;
@@ -33,7 +34,7 @@ public interface AppComponent {
 
     void inject(SearchAdapter adapter);
 
-    void inject(BookmarksDialogBuilder builder);
+    void inject(LightningDialogBuilder builder);
 
     void inject(BookmarkPage bookmarkPage);
 
@@ -46,6 +47,8 @@ public interface AppComponent {
     BookmarkPage getBookmarkPage();
 
     Bus getBus();
+
+    HistoryDatabase getHistoryDatabase();
 
     Context getApplicationContext();
