+++ app/src/main/java/acr/browser/lightning/dialog/LightningDialogBuilder.java
@@ -1,8 +1,13 @@
 package acr.browser.lightning.dialog;
 
+import android.app.Activity;
+import android.content.ClipData;
+import android.content.ClipboardManager;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.net.Uri;
+import android.os.Build;
+import android.support.annotation.NonNull;
 import android.support.v7.app.AlertDialog;
 import android.view.View;
 import android.widget.ArrayAdapter;
@@ -19,24 +24,32 @@ import javax.inject.Inject;
 import acr.browser.lightning.R;
 import acr.browser.lightning.app.BrowserApp;
 import acr.browser.lightning.bus.BookmarkEvents;
+import acr.browser.lightning.bus.BrowserEvents;
 import acr.browser.lightning.constant.Constants;
+import acr.browser.lightning.constant.HistoryPage;
 import acr.browser.lightning.database.BookmarkManager;
+import acr.browser.lightning.database.HistoryDatabase;
 import acr.browser.lightning.database.HistoryItem;
 import acr.browser.lightning.utils.Utils;
 
 /**
+ * TODO Rename this class it doesn't build dialogs only for bookmarks
+ *
  * Created by Stefano Pacifici on 02/09/15, based on Anthony C. Restaino's code.
  */
-public class BookmarksDialogBuilder {
+public class LightningDialogBuilder {
 
     @Inject
     BookmarkManager bookmarkManager;
 
     @Inject
+    HistoryDatabase mHistoryDatabase;
+
+    @Inject
     Bus eventBus;
 
     @Inject
-    public BookmarksDialogBuilder() {
+    public LightningDialogBuilder() {
         BrowserApp.getAppComponent().inject(this);
     }
 
@@ -46,7 +59,7 @@ public class BookmarksDialogBuilder {
      * @param context   used to show the dialog
      * @param url   the long pressed url
      */
-    public void showLongPressedDialogForUrl(final Context context, final String url) {
+    public void showLongPressedDialogForBookmarkUrl(final Context context, final String url) {
         final HistoryItem item;
         if (url.startsWith(Constants.FILE) && url.endsWith(Constants.BOOKMARKS_FILENAME)) {
             // TODO hacky, make a better bookmark mechanism in the future
@@ -65,19 +78,19 @@ public class BookmarksDialogBuilder {
             if (item.isFolder()) {
                 showBookmarkFolderLongPressedDialog(context, item);
             } else {
-                showLongPressedDialogForUrl(context, item);
+                showLongPressedDialogForBookmarkUrl(context, item);
             }
         }
     }
 
-    public void showLongPressedDialogForUrl(final Context context, final HistoryItem item) {
+    public void showLongPressedDialogForBookmarkUrl(final Context context, final HistoryItem item) {
         final DialogInterface.OnClickListener dialogClickListener =
                 new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
                         switch (which) {
                             case DialogInterface.BUTTON_POSITIVE:
-                                eventBus.post(new BookmarkEvents.AsNewTab(item));
+                                eventBus.post(new BrowserEvents.OpenUrlInNewTab(item.getUrl()));
                                 break;
                             case DialogInterface.BUTTON_NEGATIVE:
                                 if (bookmarkManager.deleteBookmark(item)) {
@@ -197,4 +210,103 @@ public class BookmarksDialogBuilder {
                 });
         editFolderDialog.show();
     }
+
+    public void showLongPressedHistoryLinkDialog(final Context context, final String url) {
+        DialogInterface.OnClickListener dialogClickListener = new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                switch (which) {
+                    case DialogInterface.BUTTON_POSITIVE:
+                        eventBus.post(new BrowserEvents.OpenUrlInNewTab(url));
+                        break;
+                    case DialogInterface.BUTTON_NEGATIVE:
+                        mHistoryDatabase.deleteHistoryItem(url);
+                        // openHistory();
+                        eventBus.post(new BrowserEvents.OpenUrlInCurrentTab(HistoryPage.getHistoryPage(context)));
+                        break;
+                    case DialogInterface.BUTTON_NEUTRAL:
+                        eventBus.post(new BrowserEvents.OpenUrlInCurrentTab(url));
+                        break;
+                    default:
+                        break;
+                }
+            }
+        };
+
+        AlertDialog.Builder builder = new AlertDialog.Builder(context);
+        builder.setTitle(R.string.action_history)
+                .setMessage(R.string.dialog_history_long_press)
+                .setCancelable(true)
+                .setPositiveButton(R.string.action_new_tab, dialogClickListener)
+                .setNegativeButton(R.string.action_delete, dialogClickListener)
+                .setNeutralButton(R.string.action_open, dialogClickListener)
+                .show();
+    }
+
+    // TODO There should be a way in which we do not need an activity reference to dowload a file
+    public void showLongPressImageDialog(@NonNull final Activity activity, @NonNull final String url,
+                                          @NonNull final String userAgent) {
+        DialogInterface.OnClickListener dialogClickListener = new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                switch (which) {
+                    case DialogInterface.BUTTON_POSITIVE:
+                        eventBus.post(new BrowserEvents.OpenUrlInNewTab(url));
+                        break;
+                    case DialogInterface.BUTTON_NEGATIVE:
+                        eventBus.post(new BrowserEvents.OpenUrlInCurrentTab(url));
+                        break;
+                    case DialogInterface.BUTTON_NEUTRAL:
+                        if (Build.VERSION.SDK_INT > 8) {
+                            // Should be better to send an event on the bus here
+                            Utils.downloadFile(activity, url,
+                                    userAgent, "attachment");
+                        }
+                        break;
+                }
+            }
+        };
+
+        AlertDialog.Builder builder = new AlertDialog.Builder(activity);
+        builder.setTitle(url.replace(Constants.HTTP, ""))
+                .setCancelable(true)
+                .setMessage(R.string.dialog_image)
+                .setPositiveButton(R.string.action_new_tab, dialogClickListener)
+                .setNegativeButton(R.string.action_open, dialogClickListener)
+                .setNeutralButton(R.string.action_download, dialogClickListener)
+                .show();
+    }
+
+    public void showLongPressLinkDialog(final Context context, final String url) {
+        DialogInterface.OnClickListener dialogClickListener = new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                switch (which) {
+                    case DialogInterface.BUTTON_POSITIVE:
+                        eventBus.post(new BrowserEvents.OpenUrlInNewTab(url));
+                        break;
+
+                    case DialogInterface.BUTTON_NEGATIVE:
+                        eventBus.post(new BrowserEvents.OpenUrlInCurrentTab(url));
+                        break;
+
+                    case DialogInterface.BUTTON_NEUTRAL:
+                        ClipboardManager clipboard = (ClipboardManager) context.getSystemService(Context.CLIPBOARD_SERVICE);
+                        ClipData clip = ClipData.newPlainText("label", url);
+                        clipboard.setPrimaryClip(clip);
+                        break;
+                }
+            }
+        };
+
+        AlertDialog.Builder builder = new AlertDialog.Builder(context); // dialog
+        builder.setTitle(url)
+                .setCancelable(true)
+                .setMessage(R.string.dialog_link)
+                .setPositiveButton(R.string.action_new_tab, dialogClickListener)
+                .setNegativeButton(R.string.action_open, dialogClickListener)
+                .setNeutralButton(R.string.action_copy, dialogClickListener)
+                .show();
+    }
+
 }
