+++ app/src/main/java/acr/browser/lightning/utils/WebUtils.java
@@ -10,6 +10,7 @@ import android.webkit.WebStorage;
 import android.webkit.WebView;
 import android.webkit.WebViewDatabase;
 
+import acr.browser.lightning.app.BrowserApp;
 import acr.browser.lightning.database.HistoryDatabase;
 
 /**
@@ -32,7 +33,7 @@ public class WebUtils {
     }
 
     public static void clearHistory(@NonNull Context context) {
-        HistoryDatabase.getInstance().deleteHistory();
+        BrowserApp.getAppComponent().getHistoryDatabase().deleteHistory();
         WebViewDatabase m = WebViewDatabase.getInstance(context);
         m.clearFormData();
         m.clearHttpAuthUsernamePassword();
